refactor: rename `data` folder into `const`
