+++ lib/src/domain/device_repository.dart
@@ -1,0 +1,4 @@
+
+abstract class DeviceRepository {
+  Future<Map<String, dynamic>> getDeviceLanguage();
+}
